fix bandwidth benchmark to float division

+++ src/benchmarks/shared/bandwidth/l2ReadBandwidth.cpp
@@ -82,7 +82,7 @@ double l2ReadBandwidthLauncher(size_t arraySizeBytes) {
 
 namespace benchmark {
     double measureL2ReadBandwidth(size_t l2SizeBytes) {
-        double testSizeGiB = (double)l2SizeBytes / (1 * GiB); // Convert to GiB
+        double testSizeGiB = (double)l2SizeBytes / (double)(1 * GiB); // Convert to GiB
 
         std::vector<double> results(ROUNDS);
         for (uint32_t i = 0; i < ROUNDS; ++i) {
